consolidate TokenManager into Credentials
