--- Scripts/Utilities/TokenManager.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: d25b89e13eef980438578844bd0ef760
-timeCreated: 1524510898
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
